+++ guillotina/contrib/dbusers/interfaces.py
@@ -1,0 +1,5 @@
+from zope.interface import Interface
+
+
+class IDBUsersLayer(Interface):
+    """Marker interface layer for db users"""