+++ guillotina/tests/fixtures.py
@@ -1,3 +1,4 @@
+import json
 from aiohttp.client_exceptions import ContentTypeError
 from aiohttp.test_utils import TestServer
 from guillotina import task_vars
@@ -463,3 +464,17 @@ def redis_container():
     yield host, port  # provide the fixture value
     pytest_docker_fixtures.redis_image.stop()
     annotations["redis"] = None
+
+
+class DBUsersRequester(ContainerRequesterAsyncContextManager):
+    async def __aenter__(self):
+        requester = await super().__aenter__()
+        await requester(
+            "POST", "/db/guillotina/@addons", data=json.dumps({"id": "dbusers"})
+        )
+        return requester
+
+
+@pytest.fixture(scope="function")
+async def dbusers_requester(guillotina):
+    return DBUsersRequester(guillotina)